+++ plugin/src/main/kotlin/org/jlleitschuh/gradle/ktlint/worker/KtLintWorkAction.kt
@@ -6,6 +6,7 @@ import com.pinterest.ktlint.core.ParseException
 import com.pinterest.ktlint.core.RuleSet
 import com.pinterest.ktlint.core.RuleSetProvider
 import net.swiftzer.semver.SemVer
+import org.apache.commons.io.input.MessageDigestCalculatingInputStream
 import org.gradle.api.GradleException
 import org.gradle.api.file.ConfigurableFileCollection
 import org.gradle.api.file.RegularFileProperty
@@ -14,6 +15,15 @@ import org.gradle.api.provider.Property
 import org.gradle.api.provider.SetProperty
 import org.gradle.workers.WorkAction
 import org.gradle.workers.WorkParameters
+import org.jlleitschuh.gradle.ktlint.worker.KtLintWorkAction.FormatTaskSnapshot.Companion.contentHash
+import java.io.BufferedInputStream
+import java.io.BufferedOutputStream
+import java.io.File
+import java.io.FileInputStream
+import java.io.FileOutputStream
+import java.io.ObjectInputStream
+import java.io.ObjectOutputStream
+import java.io.Serializable
 import java.util.ServiceLoader
 
 @Suppress("UnstableApiUsage")
@@ -39,6 +49,7 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
         resetEditorconfigCache()
 
         val result = mutableListOf<LintErrorResult>()
+        val formattedFiles = mutableMapOf<File, ByteArray>()
 
         parameters.filesToLint.files.forEach {
             val errors = mutableListOf<Pair<LintError, Boolean>>()
@@ -61,6 +72,7 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
                     val updatedFileContent = KtLint.format(ktLintParameters)
 
                     if (updatedFileContent != currentFileContent) {
+                        formattedFiles[it] = contentHash(it)
                         it.writeText(updatedFileContent)
                     }
                 } else {
@@ -89,6 +101,13 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
                 result,
                 parameters.discoveredErrorsFile.asFile.get()
             )
+
+        if (formattedFiles.isNotEmpty()) {
+            val snapshotFile = parameters.formatSnapshot.get().asFile
+                .also { if (!it.exists()) it.createNewFile() }
+            val snapshot = FormatTaskSnapshot(formattedFiles)
+            FormatTaskSnapshot.writeIntoFile(snapshotFile, snapshot)
+        }
     }
 
     private fun resetEditorconfigCache() {
@@ -141,5 +160,38 @@ abstract class KtLintWorkAction : WorkAction<KtLintWorkAction.KtLintWorkParamete
         val discoveredErrorsFile: RegularFileProperty
         val ktLintVersion: Property<String>
         val editorconfigFilesWereChanged: Property<Boolean>
+        val formatSnapshot: RegularFileProperty
+    }
+
+    /**
+     * Represents pre-formatted files snapshot (file + it contents hash).
+     */
+    internal class FormatTaskSnapshot(
+        val formattedSources: Map<File, ByteArray>
+    ) : Serializable {
+        companion object {
+            private const val serialVersionUID = 1L
+
+            fun readFromFile(snapshotFile: File) =
+                ObjectInputStream(BufferedInputStream(FileInputStream(snapshotFile)))
+                    .use {
+                        it.readObject() as FormatTaskSnapshot
+                    }
+
+            fun writeIntoFile(
+                snapshotFile: File,
+                formatSnapshot: FormatTaskSnapshot
+            ) = ObjectOutputStream(BufferedOutputStream(FileOutputStream(snapshotFile)))
+                .use {
+                    it.writeObject(formatSnapshot)
+                }
+
+            fun contentHash(file: File): ByteArray {
+                return MessageDigestCalculatingInputStream(BufferedInputStream(FileInputStream(file))).use {
+                    it.readBytes()
+                    it.messageDigest.digest()
+                }
+            }
+        }
     }
 }