+++ plugin/src/test/kotlin/org/jlleitschuh/gradle/ktlint/KtLintSupportedVersionsTest.kt
@@ -106,7 +106,9 @@ class KtLintSupportedVersionsTest : AbstractPluginTest() {
             "0.38.1",
             "0.39.0",
             "0.40.0",
-            "0.41.0"
+            "0.41.0",
+            "0.42.0",
+            "0.42.1"
         ).also {
             // "0.37.0" is failing on Windows machines that is fixed in the next version
             if (!OS.WINDOWS.isCurrentOs) it.add("0.37.0")